oracle/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
alembic
python-dotenv
requests
redis
pytest
pydantic-settings
